Remove missed replacement legacy import code
https://eaflood.atlassian.net/browse/WATER-4535

In [Remove the replacement legacy import code](https://github.com/DEFRA/water-abstraction-system/pull/1585) we removed the code we added for replacing the legacy import service (see the PR for the reasons why).

Or we thought we did!

Turns out we missed a bunch of stuff. Doh!

This change removes the stuff we missed.