Replace DetermineLicenceEndDateChanged service
By including both the NALD and WRLS dates in our original fetch service we can save an additional 74K queries against the database.

What our job needs to know is _if_ a licence has a changed end date. Billing flags and return logs then need to know _when_. Currently, we are passing the raw data down and they are doing there own variation of determining what the `changeDate` should be. However, its the same for both, so why not determine that in the service that is already processing the end dates!?

The result is a new DetermineEarliestLicenceChangedDate service that replaces DetermineLicenceEndDateChanged.