Add new service to process the licence
With the name change and use of DetermineEarliestLicenceChangedDateService it was easier to create a new service to handle the processing of each licence by the job.

Another feature that was missed was the inclusion of the feature flag. Until WRLS takes over management of requirements for returns, we shouldn't be reissuing return logs during import. This new service gives us an opportunity to add the functionality in.