Add new main service for the job
We needed to change the service name to follow our convention of the service matching with the job endpoint and name.

We also didn't see the need to solve the problem of importing an ESM only dependency any different than we already do with Got in our `BaseRequest` module.

That, plus some expansion of comments and tidying up resulted in this new service to replace the existing one.