Rename Import job as Licence Changes and refactor (#1593)
https://eaflood.atlassian.net/browse/WATER-4728

After completing [Remove the replacement legacy import code](https://github.com/DEFRA/water-abstraction-system/pull/1585) we could review the replacement 'import job' we created in a clearer light.

The immediate thing that rang out was that we are not importing a license! The job is connected to the import licence process, but specifically, it cares about changes to the licence. Hence the name change.

After that, we could see the job and the two downstream processes it kicks off from a bird's eye view.

What jumped out is they were both doing there own variation of the same thing: working out what the 'change date' is. Plus, in the case of return logs it was re-quering the DB for information that could be retrieved in the job's `FetchLicenceService`.

So, in this change we have refactored the job to handle both

- determining that a licence has a changed 'end date'
- determining what that changed date is

As the job has this information, it can then pass it to the downstream services, which means we can delete the redundant code from them.

Along the way, we do some housekeeping: updating the job to match the existing jobs, enhancing the comments, and fixing a whoopsie in the query `FetchLicencesServices` uses! 😱 😁