Refactor NALD query to fetch licences first
Not sure why we were selecting from `NALD_ABS_LIC_VERSIONS` and then joining to `NALD_ABS_LICENCES` but as the service is called `FetchLicencesService`, and all the data is coming from `NALD_ABS_LICENCES` it makes sense to refactor the query to treat it as the main table to select from.

Now it reads as expected, select all licences with a 'CURR' licence version.