Add guidance view on entering multiple readings
https://eaflood.atlassian.net/browse/WATER-4812

For some returns, users are required to submit readings for each day, which can be very tedious to enter manually. To simplify this, a previous page for entering multiple readings is being developed - [Enter multiple abstraction values](https://eaflood.atlassian.net/browse/WATER-4811).
Since the process requires the readings to be submitted in a specific format (either comma-separated or line-separated), some users who are unfamiliar with the process may need guidance. This will be provided through the new guidance page, which explains how to submit multiple returns on the `Enter multiple daily volumes` page.

This PR focuses on creating the static guidance page for entering multiple abstraction values.