Return Logs setup received date (#1602)
https://eaflood.atlassian.net/browse/WATER-4844

As part of the work to migrate the return logs setup journey, this PR is focused on adding the 'When was this return received' page.

NOTE: This work is only for a due return. The pages will be updated in a separate PR to handle the editing of existing return logs.