Create quarterly return logs (#1603)
https://eaflood.atlassian.net/browse/WATER-4841

As part of our work to replace NALD for the generation of return logs we're looking at introducing the concept of a quarterly return. This is an all year cycle that has been split into 4 3 month periods that water companies will return against through out the year.

This PR checks if the return version has been marked for quarterly returns, if it starts after the go live date and then creates the smaller return logs.

There is also some refactoring of the quarter period naming.