Return Logs setup - Meter provided
https://eaflood.atlassian.net/browse/WATER-4846

As part of the work to migrate the return logs setup journey, this PR is focused on adding the 'Have meter details been provided' page.

NOTE: This work is only for a due return. The pages will be updated in a separate PR to handle the editing of existing return logs.