Return Logs setup reported page (#1607)
https://eaflood.atlassian.net/browse/WATER-4808

As part of the work to migrate the return logs setup journey, this PR is focused on adding the 'How was this return reported' page.

NOTE: This work is only for a due return. The pages will be updated in a separate PR to handle the editing of existing return logs.