Return Logs setup units page (#1612)
https://eaflood.atlassian.net/browse/WATER-4810

As part of the work to migrate the return logs setup journey, this PR is focused on adding the 'Which units were used?' page.

NOTE: This work is only for a due return. The pages will be updated in a separate PR to handle the editing of existing return logs.