Update controllers to new pageTitle and navBar convention (#1620)
DEFRA/water-abstraction-team#132

During the migration of legacy pages into our new system repository, we identified inconsistencies in how the pageTitle and navBar were being managed. In some cases, they were defined in the controller, in others within the services, and occasionally in the presenter. After reviewing our coding conventions, we have agreed that, moving forward, the pageTitle should be set in the presenter, while the navBar should be managed in the service called from the controller.

This PR aligns the repository with these updated conventions.