Tidy up return logs setup journey
The first few pages for the return logs setup journey have been completed. As a result of the work for this and the peer review efforts made we have new coding conventions created and consistency issues spotted amongst the code base. This PR is addressing the presenters that use session data. We have currently been called session.data within the presenters where this is not needed. This PR brings all the pages inline with the convention of calling the property needed straight from the session rather than going through the data property.