Fix when the "Record receipt" option is displayed https://eaflood.atlassian.net/browse/WATER-4836
The "Record receipt" option is dynamic and only displayed if the return log has no receipt date recorded against it.

Currently in order to determine if the option should be displayed the session data is checked to see if the `receivedDate` date has been populated. If not the option is displayed.

The problem arises if the user goes down the Enter & Submit route, enters & submits a received date and then decides they have made an error and goes "back". In this scenario the "Record receipt" option will be hidden as the session data will now contain a `receivedDate`.

To resolve this issue when the session is initiated a new `beenReceived` data item will be added to the session data. This will be boolean set to `true` if it has been received, false if not.