Tidy up return logs setup journey (#1626)
The first few pages for the return logs setup journey have been completed. As a result of the work for this and the peer review efforts made we have new coding conventions created and consistency issues spotted amongst the code base. This PR addresses the presenters who use session data. We have currently been called session.data within the presenters where this is not needed.

This PR brings all the pages in line with the convention of calling the property needed straight from the session rather than going through the data property.