Rearrange return log setup routes (#1635)
https://eaflood.atlassian.net/browse/WATER-4836
https://eaflood.atlassian.net/browse/WATER-4844

After starting work on the return log setup journey, the pages for "What do you want to do with this return?" and "When was the return received?" were completed. Since then, the business has decided to swap the order of these pages. The journey now begins with entering the return received date, followed by selecting what to do with the return. This change makes sense, as all the options for handling the return previously led to entering a received date. By starting with the received date page, we capture that data upfront before users proceed with their chosen path. This PR implements the change to reorder these pages.