Stop displaying supp. match result in /check page
https://eaflood.atlassian.net/browse/WATER-4877

Our users have spotted an issues with the bill run setup `check` page. In general, the idea is if an existing bill run for the selected region exists, the users is blocked from creating a new one and we display the match on the `/check` page.

Supplementary is more complex, because of the two different charge schemes involved. This means we need to trigger two separate bill runs. If we find an in-progress match for both, then the user is still blocked and we display the first match.

However, if only one of the bill runs is blocked (PRESROC or SROC), then we still allow the use to create the other.

That bit works (they can still click the create button). The bug is we are displaying the in-progress match we found, rather than replaying the options selected. This makes the page confusing, because it is saying 'go ahead' whilst displaying what in other scenarios would be the blocking bill run.

This page fixes the page to handle the scenario of a match existing, but the user still being permitted to continue.