Replace `==` in the views with `===` (#1636)
It has been spotted in a few views that we have been using `==` rather than `===`.

Generally, we use the strict comparison `===` to compare content & type unless it is necessary to perform a loose equality comparison.

It looks like it is not necessary to use `==` in the views. So where applicable they will be updated to `===`.