Handle edge case in non-prod supplementary
When creating test scenarios for the primary issue came across an edge case that would only happen in non-production environments. As part of the original intent was to stop any more errors happening with this process, irrespective of environment, we update the service to cater for it.

If the last annual bill run for a region was PRESROC, when we trigger the supplementary, we should only be triggering a PRESROC bill run.