Add tests for all scenarios, including the fails
In doing this we realised we didn't really have one place that listed out the scenarios we are having to deal with. Yes, we were testing individual properties of the `CheckPresenter`, but those weren't covering all scenarios hence this issue we are now fixing.

We believe the `CheckPresenter` in its current form will become to complex if we try and refactor it to handle both when a bill run is blocked or when it is allowed to be created.

So, ahead of splitting up the functionality we add tests for each scenario we can think of to the `CheckService`, is going forward it will decide which presenter to use fo a given situation. It also gives us a handy place to record all the scenarios.