Save the region name in the session
The current CheckPresenter is the only presenter we have that has to interact with the DB. This is because we were only passing in a region ID from the session.

If the bill run is allowed to be created, we needed to fetch the region display name in order to display it on the review screen.

After looking at this again, we spotted that we fetch the display names for all regions as part of the `SubmitRegionService` (needed to validate the payload contains a recognised region). Whilst we have the information, we could persist it to the session allowing us to remove the query in the CheckPresenter.

This change doesn't bother removing anything from the presenter, as we are just about to rewrite it. But it does update `SubmitRegionService` to save the region name to the session ready for our refactor.