Update "What do you want to do with this return?" page (#1639)
https://eaflood.atlassian.net/browse/WATER-4836

After some discussion, the journey for submitting return data has changed. This has resulted in some changes being needed to the look of the page as well as the filename of the page and route.

Initially, this page was the first in the journey so the route/files had been called start. Now it is no longer the "start" of the journey it has been decided to rename it to "submission".