Add pagination to notification setup review
https://eaflood.atlassian.net/browse/WATER-4775

As part of the refactor of the previous notification service (in the legacy code) we have added new notification logic to the system.

This has been changed and enhanced. Part this change is the user of pagination on the notification setup review page.

As the query is quite complex it has been made as a raw query, as such we do not have access to previously built pagination logic for the database.

We have added the pagination logic into the presenter.

The presenter receives the de duplicated recipients and calculates the page.