Add failing test for formatAbstractionPeriod()
We display the abstraction period for various entities in a number of places, and will soon be displaying it as part of the new [view return logs page](https://github.com/DEFRA/water-abstraction-system/pull/1484).

If `formatAbstractionPeriod()` is updated to handle this scenario, we'll ensure we display a consistent message throughout the service.