Naughty! Return null instead of the message
It can be hard to keep track of some of the patterns we have been following as we build the new pages.

Because we keep referring back to the legacy pages, they will often skew how we intend to build our pages.

One of the patterns we have been following, when we remember, is to _always_ display a field, but leave it blank if there is no data.

The legacy code does different things in different places

- It will hide fields which contain no data
- It will display the field with a stock message

We fell into the trap of the latter, but thankfully caught in before it stuck!