Update the view return version presenter
Didn't spot at the time it was doing it's won implementation of `formatAbstractionPeriod()`. So, we update it to use the common one in `base.presenter.js`.

We also didn't spot it was using the legacy pattern, of sometimes prefixing dates with `From ...`. Again, using the common method in the base presenter helps us stick to a common pattern.