Add logging to match & allocate fetch service (#1650)
https://eaflood.atlassian.net/browse/WATER-4885

When attempting to run the North East two-part tariff bill run for 2023-24, the B&D team got an error. When we dug in, we found it was due to a licence having a return log with no abstraction period.

The return log had no abstraction period because the return requirement it was based on didn't have an abstraction period set.

Either way, it caused the bill run to error. The logs clearly show us that the query in `FetchReturnLogsForLicenceService` had failed, but they didn't tell us which licence had the problem. It took some re-engineering of this query and the one in `FetchChargeVersionsService` to find it!

If `FetchReturnLogsForLicenceService` had logged the licence ref when it errored, we would have tracked down and confirmed the licence with a problem return log sooner.

So, this change wraps the call in a try/catch and logs the problem licence ref if an error is thrown.