Notifications review contacts formatting
https://eaflood.atlassian.net/browse/WATER-4775

Previous work has been done to add the initial page, controller and router for the notification review page.

This change refactors the existing licence contact document header contact formatting into a shared presenter.

The notifications contact column will use this logic to give a consistent representation of the contact details to the user.