Fix logging error when deleting a bill run
We spotted in the logs that when we delete a bill run, we see the completion message but then an error message.

Just a bit worrying!

Thankfully, the deletion completes and it is just a mistake on our part not to return after logging the completion. Because we weren't, we rolled into trying to generate an error log message for an error that never happened!