Start splitting presenter logic - base check pres.
We are now starting to split up the presenter logic so each presenter can be focused on dealing with a specific scenario, rather than one presenter trying to handle all three.

The first thing we do is move the `_backLink()` to a base presenter. We have to generate a back link in all our scenarios, and whether the bill run is blocked or approved does not matter to the functionality. So, a base presenter seemed the best home for it.