Add presenter for the no annual bill run scenario
This is something we often face in our non-production environments. After being asked too many times to investigate issues with billing only to find it's because someone is trying to generate a supplementary for a region with no annual bill run, we now handle the scenario and display messages that clearly state this is why the bill run is blocked.