Rename 'Financial Year' to 'Financial year end'
When the issue was reported, the users queried what Financial year meant in respect of supplementary. It spans 5 years, but might only span a couple, or just one depending on the bills generated.

They were happy with our explanation that we were just repeating what you see when you view a bill run in the UI, for example, you might have bills from 2022, 2023, 2024 and 2025, but the UI will just show '2024 to 2025'.

But, as we have this opportunity, we try and make things a little less confusing by renaming the field 'Financial year end' and only showing the year end. That way, we are both accurate, and still give enough information to verify this is the bill run they intend to create.