Stop displaying supp. match result in /check page (#1638)
https://eaflood.atlassian.net/browse/WATER-4877

Our users have spotted an issue with the bill run setup `check` page. In general, the idea is that if an existing bill run for the selected region exists, the user will be blocked from creating a new one, and we will display the match on the `/check` page.

Supplementary is more complex because of the two different charge schemes involved. This means we need to trigger two separate bill runs. If we find an in-progress match for both, the user is still blocked, and we will display the first match.

However, if only one of the bill runs is blocked (PRESROC or SROC), we still allow the user to create the other.

That bit works (they can still click the create button). The bug is that we display the in-progress match we found rather than replaying the selected options. This makes the page confusing because it says 'go ahead' while displaying what, in other scenarios, would be the blocking bill run.

This page fixes the page to handle the existing match scenario, but the user is still permitted to continue.