Remove where clause on fetch return cycles (#1659)
https://eaflood.atlassian.net/browse/WATER-4546

As part of the testing of this story it was found that we were not voiding all the return logs after the licence end dates return cycle. This PR fixes that by moving the logic into the process return cycle function. When it is now looping through all the return cycles it will not generate returns for cycles after the licence end date but will continue to void them.