Add `notes` and `created_by` to `return_submissions` https://eaflood.atlassian.net/browse/WATER-4807
New functionality is being created to enable the user to save a note against a return submission, previously referred to as a version.

This PR will add the `notes` and `created_by` columns to the `returns.versions` table in our test DB as well as updating the `return_submissions` view.

PR https://github.com/DEFRA/water-abstraction-returns/pull/406 will make these updates to the `returns.versions` table.