Refactor licence contact details into CRM presenter
https://eaflood.atlassian.net/browse/WATER-4775

As part of the work to migrate the notifications into the system repo.

We currently format the contact from crm.document_header.metadata in the view licence contact details presenter.

We now have a requirement to use the contact details in multiple places.

This change refactors the code from the  view licence contact details presenter into a shared crm presenter.