Sort Notifications recipients (#1664) https://eaflood.atlassian.net/browse/WATER-4775
This change introduces a sort on the recipients array in the presenter. It sorts by the contacts name.

We need to work out the name before we can sort. So we first map the recipients and then sort.

This will result in all the recipients being mapped and sorted, with only 25 (the current default pagination per page) of the results being used. We have deemed this acceptable for the sake of simplicity and have accounted for performance concerns.