Add more contacts to licence contact details
https://eaflood.atlassian.net/browse/WATER-4879

We need to add the primary user and returns agent to the licence contact details page.

This page shows the contacts stored in crm.

Ths change will update the fetch query to return the 'primary_user' and 'user_agent' if present in the database.

These detail will be displayed in the view as their own contact block.