Create new quarterly return versions for water companies (#1648)
https://eaflood.atlassian.net/browse/WATER-4881

As part of our work in setting up quarterly returns for water companies we need to run a one off migration for each of them to create a new return version. This PR creates an end point that can be hit as part of a job. It will loop through each of the water companies that do not have a return version starting next year and create one.