Update Licence entity role to Has Many Relation
https://eaflood.atlassian.net/browse/WATER-4879

As part of the work to show the "Primary user" and "Returns agent" in the licence contact details page. We noticed our previous assumption of a licence entity role having one relation was incorrect.

A licence entity role can have many relations. This is commonly found when a licence entity role with a 'Primary user' also has a 'Returns agent' (know as a 'user'agent' in the db)

This change updates the `licenceEntityRole` to `licenceEntityRoles` and updates any necessary code / tests that fail.