Return log 'received' confirmation page
https://eaflood.atlassian.net/browse/WATER-4888

When a user has started the `Submit` journey on a `Due` return log, has completed the two previous pages, and selected `Record receipt` they will arrive
at the confirm received page. This page will update the return log for the licence with a status of `Received` and updated `received_date` in the database.

This PR creates the return log received confirmation page.