Refactor Determine Returns Period
https://eaflood.atlassian.net/browse/WATER-4776

As part of the work to implement notifications in the system repo we have found some logic that will need to be duplicated in upcoming work.

This change lifts the determine returns periods logic into a shared service.

This can / will be used in multiple places and is complex enough to justify lifting into it own service and associated tests.