Refactor existing CSV converter
https://eaflood.atlassian.net/browse/WATER-4776

As part of the work to implement notifications in the system repo we have found some logic that will need to be duplicated in upcoming work.

This change lifts the convert to csv service into the lib.

It has been renamed to transform to csv,

The existing logic has been updated to be more precise in the transformation it is doing (transforms an array into a CSV row).