Add FetchDownloadRecipientsService https://eaflood.atlassian.net/browse/WATER-4776
As part of the work to implement notifications in the system repo we have a requirement to download the recipients data as a CSV.

This change adds the required data for the CSV. This recipient list does not require deduplication as per previous work. And we can expect multiple rows per licence.

However, we will still a registered licence contact over an unregistered licence.

 The controller / presenter / service will be added in a later change.