Add 'Mark as under query' button to View Return Logs page (#1671)
https://eaflood.atlassian.net/browse/WATER-4893

This PR adds a "Mark under query" button to the "View return log" page.

If the return log is not already under query, the button text will be `Mark as under query`. Clicking it sets the `underQuery` flag on the return log to `true` and reloads the page displaying a temporary flash banner.

If the return log is currently under query, the button text will be `Resolve query`. Clicking it sets the `underQuery` flag to `false` and reloads the page. No banner will be shown in this scenario.

We introduce `SubmitViewReturnLogService` to handle this. The service handles checking if the button click action is to mark or resolve the query; creates a banner using `yar.flash()` if it's being marked; and updates the record.