Return Logs setup - Period used (#1660)
https://eaflood.atlassian.net/browse/WATER-4853

As part of the work to migrate the return logs setup journey, this PR is focused on adding the 'What period was used for this volume?' page.

NOTE: This work is only for a due return. The pages will be updated in a separate PR to handle the editing of existing return logs.