Fix return logs setup single volume validation (#1689)
https://eaflood.atlassian.net/browse/WATER-4852

While testing for the above ticket, it was noticed that the single volume page didn't allow a user to enter a number less than 1 and validate it. If a user enters 0.5, then this should be accepted. This PR updates the validation to only allow positive numbers, thus removing the need for a min volume.