Update return log setup initialise service
https://eaflood.atlassian.net/browse/WATER-4852

With the last page created for the return log setup journey, its now time to focus on the return log submission lines. A user can enter a single volume for the return and a period that the water was abstracted from and we need to work out the split. This PR is focusing on creating the lines object in the session data. This Object is created from the return log start and end date and the frequency in which the return is reported.