Make sure meter details are null if not provided Previously a situation could arise where the user has said "yes" to "Have meter details been provided?", entered some meter details, realised they have made a mistake and gone back to the previous page to answer "no" to the question.
In this situation the session data would contain `meterProvided: 'no'` but still contain the meter details that were previously entered.

This could potentially cause issues in the check details screen when "changing" the meter details. As when determining if the meter details should be presented to the user or the "Not provided" text, it is not possible to rely on the value of `meterProvided` since this isn't updated when the meter details are changed.