Add remove licences page for notifications setup (#1690)
https://eaflood.atlassian.net/browse/WATER-4777

As part of the work to implement notifications in the system repo we have a requirement allow the user to remove licences from teh recipients list.

This change adds the view element of the remove-licences page. It adds the controller, presenter and service to display the remove-licences page.

The submit and re-render logic will be done in a later change.