Add submit remove licences page for notifications setup
https://eaflood.atlassian.net/browse/WATER-4777

As part of the work to implement notifications in the system repo we have a requirement allow the user to remove licences from teh recipients list.

This change adds the basic submit logic for the notifications setup remove licences page.

The logic to calculate if a licence is valid will be done in a future change.