Add remove licences from notification validation
https://eaflood.atlassian.net/browse/WATER-4777

As part of the work to implement notifications in the system repo we have a requirement allow the user to remove licences from teh recipients list.

This change updates the validation for the remove licences page. The user should be allowed to remove all the licences. In case they have made a mistake and do not need to remove any licences.

The validation should only allow licences seperated by a comma or a new line.