update orderBy for return logs  (#1699)
https://eaflood.atlassian.net/browse/WATER-4892

As part of our work to build out the acceptance tests for the return logs journey it was found that the return logs were not consistent when displayed on the returns tab. This PR adds another clause to the order by to make each load of the page consistent.

We also update the tear down logic for the returns schema to delete the newly generated return logs as part of the tests.