Fix issue of "overlapping" months
We encountered issues where a return starts partway through a month, and ends partway through the month a year later (for example, if the first return line is dated 2 Jan 2014 and the last return line is dated 1 Jan 2015). This was caused by us only using the month to determine whether a line should be displayed. We amend this to specifically look at the year as well.